requirements.txt
asgiref==3.11.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
uritemplate==4.2.0
Faker==22.5.1
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==24.1